Refactor configuration management by renaming config_v4.py to cd_config.py and updating references in builder, my_app, and volume_setup files. This change improves clarity and consistency in configuration handling.
